move send_email to common - fixes #545
